fetch-yarn-deps: remove prefetch-yarn-deps
